pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "forthic"
version = "0.1.0"
description = "A modern Python runtime for Forthic - a stack-based, concatenative programming language"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "BSD-2-Clause" }
authors = [
    { name = "Forthic Team" }
]
keywords = ["forthic", "stack-based", "concatenative", "interpreter", "runtime"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "python-dateutil>=2.8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
pandas = [
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",  # For Excel support
]
grpc = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "pyyaml>=6.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "pandas>=2.0.0",  # Include for testing
    "openpyxl>=3.1.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
]

[project.urls]
Homepage = "https://forthic.org"
Documentation = "https://docs.forthic.org/python"
Repository = "https://github.com/forthic/forthic"

[project.scripts]
forthic-generate-docs = "scripts.generate_docs:main"
forthic-py-server = "forthic.grpc.server:main"

[tool.hatch.build.targets.wheel]
packages = ["forthic"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["forthic"]
